fix(daemon-pty): apply render-capability env layering on daemon path
#166 added a TERM / COLORTERM backstop and the layered env stack
(SHELL_SESSIONS_DISABLE → render capability → system locale → login-shell
dotfile env → caller env) to `pty::PtyManager::spawn`. The daemon-side
spawn at `daemon::pty::PtyManager::spawn` shipped raw caller env to the
child instead, so when the daemon killswitch was on the child shell
inherited whatever `TERM` the daemon process had — typically empty under
launchd — and zsh collapsed to dumb terminfo. The result was the same
pile of redraw / color regressions #166 fixed for the in-process path:
duplicated CJK syllables, ghosted history suggestions, broken Backspace
render, and color-aware CLIs falling back to plain output.

Extract the layering into a new `pty_env` module and call it from both
`pty::PtyManager::spawn` and `daemon::pty::PtyManager::spawn` so the two
paths produce identical child env regardless of which one services a
session. Move the three backstop tests with it; the rest of `pty.rs`
stays put.
